Re-engineer the Manage Courses API client to use exceptions
Use exceptions instead of statuses, as any failures in the API calls
are exceptional by nature (inconsistent data, misconfiguration of the API, network
failures) rather than incorrect actions by end-users.

Furthermore, exceptions allow packing more info about the failure, closer to the source
of the failure so the knowledge is better encapsulated in the API client.